# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartSave Student API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Expense {
  id: number;
  amount: number;
  category: string;
  description: string;
  date: string;
  /** @nullable */
  mood?: string | null;
  createdAt: string;
}

export interface ExpenseInput {
  amount: number;
  category: string;
  description: string;
  date: string;
  mood?: string;
}

export interface ExpenseUpdate {
  amount?: number;
  category?: string;
  description?: string;
  date?: string;
  mood?: string;
}

export interface Budget {
  id: number;
  category: string;
  monthlyLimit: number;
  currentSpend: number;
  month: string;
  percentUsed?: number;
  isOverBudget?: boolean;
}

export interface BudgetInput {
  category: string;
  monthlyLimit: number;
  month: string;
}

export interface BudgetUpdate {
  monthlyLimit?: number;
}

export interface SavingsGoal {
  id: number;
  name: string;
  targetAmount: number;
  currentAmount: number;
  /** @nullable */
  deadline?: string | null;
  icon?: string;
  percentComplete?: number;
  isCompleted?: boolean;
  createdAt: string;
}

export interface SavingsGoalInput {
  name: string;
  targetAmount: number;
  deadline?: string;
  icon?: string;
}

export interface SavingsGoalUpdate {
  name?: string;
  targetAmount?: number;
  deadline?: string;
  icon?: string;
}

export interface GoalContribution {
  amount: number;
}

export interface Challenge {
  id: number;
  name: string;
  description: string;
  targetAmount: number;
  currentAmount: number;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  status: string;
  xpReward?: number;
  percentComplete?: number;
  createdAt: string;
}

export interface ChallengeInput {
  name: string;
  description: string;
  targetAmount: number;
  startDate?: string;
  endDate?: string;
  xpReward?: number;
}

export interface ChallengeUpdate {
  currentAmount?: number;
  status?: string;
}

export interface Achievement {
  id: number;
  name: string;
  description: string;
  icon: string;
  isUnlocked: boolean;
  /** @nullable */
  unlockedAt?: string | null;
  xpValue?: number;
}

export interface DashboardSummary {
  totalSpentThisMonth: number;
  totalBudget: number;
  totalSaved: number;
  budgetRemaining: number;
  expenseCount: number;
  healthScore: number;
  xpTotal: number;
  level: number;
  /** @nullable */
  topCategory?: string | null;
  recentExpenses?: Expense[];
}

export interface HealthScore {
  score: number;
  savingsRate: number;
  budgetAdherence: number;
  goalAchievement: number;
  spendingConsistency: number;
  grade?: string;
  tip?: string;
}

export interface CategorySpending {
  category: string;
  amount: number;
  percentage: number;
  budget: number;
  isOverBudget?: boolean;
}

export interface MonthlyTrend {
  month: string;
  totalSpent: number;
  totalBudget: number;
  totalSaved: number;
}

export interface SpendingForecast {
  predictedMonthlySpend: number;
  currentMonthSpend: number;
  daysLeft: number;
  dailyAverage: number;
  projectedSavings: number;
  overspendingRisk?: string;
  recommendation?: string;
}

export interface Alert {
  id: string;
  type: string;
  message: string;
  severity: string;
  /** @nullable */
  category?: string | null;
}

export type ListExpensesParams = {
category?: string;
/**
 * Format: YYYY-MM
 */
month?: string;
search?: string;
};

export type ListBudgetsParams = {
/**
 * Format: YYYY-MM
 */
month?: string;
};

export type GetSpendingByCategoryParams = {
/**
 * Format: YYYY-MM
 */
month?: string;
};