# lib/api-zod/src/generated/types/achievement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartSave Student API
 * OpenAPI spec version: 0.1.0
 */

export interface Achievement {
  id: number;
  name: string;
  description: string;
  icon: string;
  isUnlocked: boolean;
  /** @nullable */
  unlockedAt?: Date | null;
  xpValue?: number;
}